pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "CLI tool: summarize PDFs into structured JSON/Markdown via an LLM"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name" }]
dependencies = [
  "pypdf>=5.0.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "openai>=1.43.0",
  "tiktoken>=0.7.0",
  "typer>=0.12.0",
  "rich>=13.7.0"
]

[project.scripts]
research-agent = "research_agent.cli:app"

[tool.ruff]
line-length = 100
